Added in task logging for other tasks.
Added in util file for handling some of the code for
getting the logging data. One example is the username.
Added new task log structures that can be serialized
to json for each of the task types.
Made it so main only logs to stdout and the task files
log to tasks.log